Refactor configuration; drop sql config
This change extracts Configuration into it's own file, and makes a note of how to add additional configuration that gets synced to vscode. Adds max_connections to configuration.
