+++ ossdbtoolsservice/workspace/contracts/did_change_config_notification.py
@@ -3,108 +3,9 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-from enum import Enum
-from typing import Any
-
 from ossdbtoolsservice.hosting import IncomingMessageConfiguration
 from ossdbtoolsservice.serialization import Serializable
-from ossdbtoolsservice.utils import constants
-
-
-class SQLConfiguration(Serializable):
-    """
-    Configuration for SQL settings in general. These are common to any SQL provider
-    """
-
-    @classmethod
-    def get_child_serializable_types(cls) -> dict[str, type["IntellisenseConfiguration"]]:
-        return {"intellisense": IntellisenseConfiguration}
-
-    @classmethod
-    def ignore_extra_attributes(cls) -> bool:
-        return True
-
-    def __init__(self) -> None:
-        self.intellisense: IntellisenseConfiguration = IntellisenseConfiguration()
-
-
-class PGSQLConfiguration(Serializable):
-    """
-    Configuration for PGSQL tool service
-    """
-
-    @classmethod
-    def get_child_serializable_types(cls) -> dict[str, type["FormatterConfiguration"]]:
-        return {"format": FormatterConfiguration}
-
-    @classmethod
-    def ignore_extra_attributes(cls) -> bool:
-        return True
-
-    def __init__(self) -> None:
-        self.default_database: str = constants.DEFAULT_DB[constants.PG_DEFAULT_DB]
-        self.cosmos_default_database: str = constants.DEFAULT_DB[
-            constants.COSMOS_PG_DEFAULT_DB
-        ]
-        self.format: FormatterConfiguration = FormatterConfiguration()
-
-
-class Case(Enum):
-    """Case options for keyword and identifier formatting"""
-
-    UPPER = ("upper",)
-    LOWER = ("lower",)
-    CAPITALIZE = "capitalize"
-
-
-class FormatterConfiguration(Serializable):
-    """
-    Configuration for Formatter settings
-    """
-
-    @classmethod
-    def ignore_extra_attributes(cls) -> bool:
-        return True
-
-    def __init__(self) -> None:
-        self.keyword_case: str | None = None
-        self.identifier_case: str | None = None
-        self.strip_comments: bool = False
-        self.reindent: bool = True
-
-
-class IntellisenseConfiguration(Serializable):
-    """
-    Configuration for Intellisense settings
-    """
-
-    @classmethod
-    def ignore_extra_attributes(cls) -> bool:
-        return True
-
-    def __init__(self) -> None:
-        self.enable_intellisense = True
-        self.enable_suggestions = True
-        self.enable_lowercase_suggestions = False
-        self.enable_error_checking = True
-        self.enable_quick_info = True
-
-
-class Configuration(Serializable):
-    """
-    Configuration of the tools service
-    """
-
-    @classmethod
-    def get_child_serializable_types(cls) -> dict[str, type[Any]]:
-        return {"sql": SQLConfiguration, "pgsql": PGSQLConfiguration}
-
-    def __init__(self) -> None:
-        self.sql = SQLConfiguration()
-        self.pgsql = PGSQLConfiguration()
-
-    def get_configuration(self) -> PGSQLConfiguration:
-        return self.pgsql
+from ossdbtoolsservice.workspace.contracts.configuration import Configuration
 
 
 class DidChangeConfigurationParams(Serializable):
@@ -116,8 +17,8 @@ class DidChangeConfigurationParams(Serializable):
     def get_child_serializable_types(cls) -> dict[str, type[Configuration]]:
         return {"settings": Configuration}
 
-    def __init__(self) -> None:
-        self.settings: Configuration | None = None
+    def __init__(self, configuration: Configuration | None = None) -> None:
+        self.settings = configuration
 
 
 DID_CHANGE_CONFIG_NOTIFICATION = IncomingMessageConfiguration(
